+++ R/mod_add_metadata.R
@@ -47,7 +47,8 @@ mod_add_metadata_ui <- function(id){
             title = "Explanation",
             content = HTML(
               "Your metadata Excel file should contain a column with the sample ID's, and one more or more columns with metadata.",
-              "Each sample ID should be present only once in your Excel file",
+              "<br><br>",
+              "<strong>Each sample ID should be present only once in your Excel file</strong>",
               "(even if it is present multiple times in your plate design)."
             ),
             trigger = "hover",